feat: split billableMetricSetup into small files
